chore(azure_blob): split module in files
